refactor(messagecontrol): handle mutations in hook
